refactor: move analysis and dataset functions to own scripts
