+++ preprocess/text.py
@@ -1,22 +1,11 @@
-# Inspired of https://fairyonice.github.io/Develop_an_image_captioning_deep_learning_model_using_Flickr_8K_data.html
 import string
 from collections import Counter
 
-import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
 import sentencepiece as spm
 from keras.preprocessing.sequence import pad_sequences
 from tqdm import tqdm
-
-# Path to captions file
-# captions_path = "C:/Users/User/Downloads/Flickr8k_text/Flickr8k.lemma.token.txt"
-
-captions_path = "C:/Users/kaspe/Documents/Study/Q3 Deep Learning/deeplearning/dataset/Flickr8k/Flickr8k_text/Flickr8k.lemma.token.txt"
-
-# For analysis
-analysis = False
-topn = 50
 
 
 def read_file(path):
@@ -51,19 +40,7 @@ def split_lines(text):
     return datatxt
 
 
-def set_seed(sd=42):
-    from numpy.random import seed
-    from tensorflow import set_random_seed
-    import random as rn
-    # numpy random seed
-    seed(sd)
-    # core python's random number
-    rn.seed(sd)
-    # tensor flow's random number
-    set_random_seed(sd)
-
-
-def make_df_word(df_txt):
+def make_df_word(df_txt, analysis=True):
     """"
     Creates a data frame for the words
     """
@@ -80,15 +57,6 @@ def make_df_word(df_txt):
     dfword = dfword.sort_values("count", ascending=False)
     dfword = dfword.reset_index()[["word", "count"]]
     return dfword
-
-
-def plot_hist(dfsub, title="Give a title next time"):
-    plt.figure(figsize=(20, 3))
-    plt.bar(dfsub.index, dfsub["count"])
-    plt.yticks(fontsize=20)
-    plt.xticks(dfsub.index, dfsub["word"], rotation=90, fontsize=20)
-    plt.title(title, fontsize=20)
-    plt.show()
 
 
 def remove_punctuation(text_original):
@@ -127,36 +95,12 @@ def text_clean(text_original):
     return text
 
 
-def print_progress_bar(iteration, total, prefix='', suffix='', decimals=1, length=100, fill='█'):
-    """
-    Retrieved from https://stackoverflow.com/questions/3173320/text-progress-bar-in-the-console
-
-    Call in a loop to create terminal progress bar
-    @params:
-        iteration   - Required  : current iteration (Int)
-        total       - Required  : total iterations (Int)
-        prefix      - Optional  : prefix string (Str)
-        suffix      - Optional  : suffix string (Str)
-        decimals    - Optional  : positive number of decimals in percent complete (Int)
-        length      - Optional  : character length of bar (Int)
-        fill        - Optional  : bar fill character (Str)
-    """
-    percent = ("{0:." + str(decimals) + "f}").format(100 * (iteration / float(total)))
-    filledLength = int(length * iteration // total)
-    bar = fill * filledLength + '-' * (length - filledLength)
-    print('\r%s |%s| %s%% %s' % (prefix, bar, percent, suffix), end='\r')
-    # Print New Line on Complete
-    if iteration == total:
-        print()
-
-
 def read_captions(captions_path):
     text = read_file(captions_path)
     text = split_lines(text)
 
     # Create Dataframe with pandas
     df_txt = pd.DataFrame(text, columns=["image_idx", "caption_idx", "caption"])
-    # df_word = make_df_word(df_txt)
 
     return df_txt
 
@@ -210,8 +154,9 @@ def encode_caption_as_ids(df_txt, sp):
 
 
 def main2():
+    """ Steps to process text -- is implemented in the prepo.py file"""
     vocab = 2048
-    df_txt = read_captions(captions_path)
+    df_txt = read_captions("")
     df_txt = clean_captions(df_txt)
 
     # Save all of the captions as .txt for SP
@@ -224,87 +169,3 @@ def main2():
     # Whoops this pickle is to big
     # result = onehot_encode(df_txt, sp, vocab)
     # result.to_pickle('encoded_captions.p')
-
-
-if __name__ == '__main__':
-    main2()
-    # Configure details
-    # warnings.filterwarnings("ignore")
-    # config = tf.ConfigProto()
-    # config.gpu_options.per_process_gpu_memory_fraction = 0.95
-    # config.gpu_options.visible_device_list = "0"
-    #
-    # # Start tf session
-    # set_session(tf.Session(config=config))
-    #
-    # # Read file
-    # print("Start reading file\n")
-    # text = read_file(captions_path)
-    # text = split_lines(text)
-    #
-    # # Create Dataframe with pandas
-    # df_txt = pd.DataFrame(text, columns=["image_idx", "caption_idx", "caption"])
-    # df_word = make_df_word(df_txt)
-    #
-    # if analysis:
-    #     print("Start file analysis\n")
-    #
-    #     # Find unique image_idx
-    #     uni_image_idx = np.unique(df_txt.image_idx.values)
-    #     print("The number of unique file names : {}".format(len(uni_image_idx)))
-    #     print("The distribution of the number of captions for each image:")
-    #     print(Counter(Counter(df_txt.image_idx.values).values()))
-    #     print()
-    #
-    #     # Word analysis
-    #     print("Most occuring words\n")
-    #     print(df_word.head(5))
-    #
-    #     plot_hist(df_word.iloc[:topn, :], title="The top " + str(topn) + " most frequently appearing words")
-    #     plot_hist(df_word.iloc[-topn:, :], title="The least " + str(topn) + " most frequently appearing words")
-    #
-    # # Clean text
-    # print("Start cleaning captions")
-    # l = len(df_txt.caption)
-    # print_progress_bar(0, l, prefix='Progress:', suffix='Complete', length=50)
-    # for i, caption in enumerate(df_txt.caption.values):
-    #     newcaption = text_clean(caption)
-    #     df_txt["caption"].iloc[i] = newcaption
-    #     print_progress_bar(i + 1, l, prefix='Progress cleaning captions:', suffix='Complete', length=50, decimals=1)
-    #
-    # # Analyze cleaned data
-    # if analysis:
-    #     dfword = make_df_word(df_txt)
-    #     plot_hist(dfword.iloc[:topn, :], title="The top " + str(topn) + " most frequently appearing words")
-    #     plot_hist(dfword.iloc[-topn:, :], title="The least " + str(topn) + " most frequently appearing words")
-    #
-    # # Save to txt for encoding
-    # np.savetxt(r'text_dataframe.txt', df_txt["caption"], fmt='%s')
-    #
-    # # Train SentencePiece Model
-    # model_prefix = "trained_sp"
-    # spm.SentencePieceTrainer.Train(f'--input=text_dataframe.txt --model_prefix={model_prefix} --vocab_size=2048 --pad_id=2 --unk_id=1 \
-    #          --bos_id=0 --eos_id=3 ')
-    #
-    # # Load trained model
-    # sp = spm.SentencePieceProcessor()
-    # sp.Load(f'{model_prefix}.model')
-    #
-    # # Encode all captions with sentencepiece
-    # print_progress_bar(0, l, prefix='Progress:', suffix='Complete', length=50)
-    # maxlen = 0
-    #
-    # sp.SetEncodeExtraOptions(extra_option='bos:eos')
-    # for i, caption in enumerate(df_txt.caption.values):
-    #     newcaption = sp.EncodeAsIds(caption)
-    #     df_txt["caption"].iloc[i] = newcaption
-    #     if maxlen < len(newcaption):
-    #         maxlen = len(newcaption)
-    #     print_progress_bar(i + 1, l, prefix='Progress enconding captions:', suffix='Complete', length=50, decimals=1)
-    #
-    # for i, caption in enumerate(df_txt.caption.values):
-    #     padded = pad_sequences([caption], maxlen=maxlen, value=2, padding='post').flatten()
-    #     df_txt["caption"].iloc[i] = to_categorical(padded, num_classes=2048)
-    #
-    # # Write to pickle file
-    # df_txt.to_pickle(path="encoded_captionsencoded_captions.p")
